etl/requirements.txt
elasticsearch[async]==8.5.2
pydantic==1.10.2
redis==4.3.5
types-redis==4.3.21.6
psycopg2-binary==2.9.4
types-psycopg2==2.9.21.1
